requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
pymysql
python-jose[cryptography]
passlib[bcrypt]
email-validator
bcrypt==4.0.1